+++ crates/crb-agent/src/performers/mod.rs
@@ -1,7 +1,7 @@
 pub mod async_performer;
 pub mod consume_performer;
+pub mod in_context_performer;
 pub mod interrupt_performer;
-pub mod loopback;
 pub mod process_performer;
 
 #[cfg(feature = "sync")]